+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 8.4.0-dev
 - allowed setting base url for Nette\Application\LinkGenerator from Nette DIC extension
 - added arguments --bootstrap and --list-test-suites for mytester.php
 - allowed using methods without test prefix as test methods in TestCase
+- added options coverage and results for Nette DIC extension
 
 Version 8.3.0
 - added arguments --coverage and --results for mytester.php